Simplify the history extension config
All configurations (including the implementation class and jdbc connection configurations) are in one same place (a properties file).
Rely on a jvm property (extension.query.history.confi) to find the properties file and rely on the file to create the extension, disable extension if there is a failure during the process or the property is not defined.

Provide a readme to explain the usage.